+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to
 
 ## [Unreleased]
 
+- Implement `AsyncIterable` on the `MatcherStream` class, enabling clean consumption of stream events (both `dirent` and `end`) via `for await...of` loops.
 - Rename `Rule.pattern`, `PatternCache.patternContext` to `list`.
 - Add `CustomRule`. Split `Rule` into `GlobRule` and `CustomRule`.
 - Fix `match.ignored`. Were always false.